Buat fitur pemesanan produk dan perbaikan UI
Perubahan ini mencakup penambahan fungsionalitas untuk memesan produk dari marketplace, termasuk pengambilan detail produk secara *real-time* (stok dan harga), alur checkout, dan penempatan pesanan. Selain itu, dilakukan beberapa perbaikan kecil pada antarmuka pengguna (UI) di berbagai layar.

**Rincian Perubahan:**

*   **Fitur Pemesanan Produk:**
    *   Membuat `ProductDetailProvider` untuk mengelola state dan logika pengambilan data detail produk serta proses pemesanan.
    *   Mengimplementasikan pengambilan data stok dan harga produk secara *real-time* di halaman detail produk.
    *   Mengintegrasikan alur pemesanan mulai dari halaman detail, checkout, hingga konfirmasi pesanan.
    *   Menambahkan service baru di `MarketplaceService` untuk `fetchProductDetail` dan `placeOrder`.
*   **Perbaikan UI:**
    *   Mengganti beberapa widget `CustomText` dengan `Text` standar untuk konsistensi.
    *   Menghilangkan properti `fontSize` yang tidak diperlukan pada beberapa `CustomText`.
    *   Menambahkan transisi `Hero` pada gambar di `GiftCard`.
    *   Memperbaiki warna teks pada tombol di halaman "Mulai Sekarang".

+++ lib/features/marketplace/widgets/product_checkout_detail.dart
@@ -17,8 +17,8 @@ class ProductCheckoutDetail extends StatelessWidget {
     required this.onDecrease,
   });
 
-  double get price => product.price;
-  double get total => price * quantity;
+  int get price => product.price;
+  int get total => price * quantity;
 
   bool get isOutOfStock => product.stock <= 0;
   bool get isMaxStock => quantity >= product.stock;
